workspace/requirements.txt
crewai>=0.28.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0
tenacity>=8.2.0
rich>=13.7.0
pytest>=7.4.0
pyyaml>=6.0.1
jinja2>=3.1.2
